requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
httpx==0.28.1
pydantic==2.10.6
gradio==4.44.1
huggingface_hub==0.25.2
python-dotenv==1.0.1
requests==2.32.3
pypdf==5.2.0
python-docx==1.1.2
